Use (less) keyword to import CSS files instead of (import)
Apparently, less doesn't include `(import)`ed files in the dependency graph that's generated when files are built. This is a problem for less-middleware, because that module has its own cache for imported files and won't trigger a rebuild unless it sees that imported modules have been updated. 

Effectively, one has to restart the whole app right now to get CSS files that are watched with a separate module to be loaded :( With this change, that should no longer be necessary!