2.103.0:
* Frustrations with conflict resolution have been much improved. First, Apostrophe no longer displays the "another user has taken control of the document" message multiple times in a row. Second, due to changes in what browsers allow to happen when you leave the page, beginning in version 2.102.0 Apostrophe displayed too many messages about a conflict with your **own** work in another tab. We no l
onger display these messages. However, if there really *is* work lost for the sa
me document in another tab, Apostrophe will still tell you what happened in orde
r to teach the habit of not editing the same page in two tabs simultaneously.
* You may now use `select` schema field with dynamic choices as a filter in "Manage Pieces."
* `required` is now enforced on the server side for `joinByOne`. However, note t
hat it is always possible for the document you joined with to be moved to the tr
ash at a later time. You must therefore always check that the join was really lo
aded before relying on it. Thanks to Ricardo José Rodríguez Álvarez.
* Hidden information at the bottom of certain modals has been restored to view.