Pass className option to rich text widget editor
This updates the rich text editor widget to also respect the "className"
option set via widget options on an area field.

The "className" option allows to set classes on core widgets. It can be
set on the widget module or on the widget options in an area field. The
value set via the widget options in an area field overrides an already
set value via the widget module options.

See: https://v3.docs.apostrophecms.org/guide/core-widgets.html#setting-a-css-class-on-core-widgets
for reference.

While this already works in the "normal" presentation mode, in edit mode
the rich text widget was still always using the "className" set on the
widget module, even if there would be an override in the respective area
options.

With this change, the classes are applied both, in presentation as well
as in edit mode.